remove auto-sync workflow
It was introduced in 4415cd052215f194b4b13f3860e9c4ab131ecc3f with the
rationale that it was needed to test the "push" trigger for the
scorecards-bash and scorecards-golang workflows. These workflows are now
known as scorecards-head and scorecards-latest-release, neither of which
have a "push" trigger anymore.
https://github.com/ossf-tests/scorecard-action/pull/4

Syncing in general is problematic and requires a personal access token,
plus it generates issues when it fails.

Fixes https://github.com/ossf/scorecard-action/issues/1405
Fixes https://github.com/ossf/scorecard-action/issues/1406

Signed-off-by: Spencer Schrock <sschrock@google.com>